refactor & leftover
