+++ src/log/index.ts
@@ -1,5 +1,5 @@
 import { KafkaRelayLogger } from './kafka-pino-adapter';
-import { pinoLogger } from './pino-pretty-logger';
+import { pinoLogger } from './pino-logger';
 
 const kafkaRelayLogger = new KafkaRelayLogger(pinoLogger);
